test: add new tests and fix previous unit tests
