+++ package.json
@@ -25,6 +25,7 @@
     "cz-conventional-changelog": "^3.3.0",
     "dotenv": "^8.2.0",
     "mocha": "^8.2.1",
+    "mock-fs": "^4.13.0",
     "prettier": "^2.2.1"
   },
   "dependencies": {
